Math Hurts
Rewrote function which checks for a cells overlap with the polygon. Prior method appeared to work due to the wide angles of the hexadecagon but shrinking to the 8 sensor polygon made innaccuracies too obvious.

Cells now all connect correctly and weights are calculated correctly.

The only thing left to implement is an output Robot class and possibly a calculation for part weights.